.2ndBrain/requirements.txt
# Second Brain - Python Dependencies

# Core transcription
git+https://github.com/m-bain/whisperx.git
torch>=2.0.0
torchaudio>=2.0.0

# Vector database & semantic search
chromadb>=0.4.0
sentence-transformers>=2.2.0

# OCR for images
pytesseract>=0.3.10
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
